moved modal from parent to child so tests will run better
